pen2grade-ai/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
pytesseract==0.3.10
Pillow==10.3.0
python-multipart==0.0.9
google-generativeai==0.4.1
